Abstract the to state path method
