+++ synthesizer/src/ledger/mod.rs
@@ -24,14 +24,22 @@ mod iterators;
 mod latest;
 
 use crate::{
-    block::{Block, BlockTree, Header, HeaderLeaf, Metadata, Origin, Transaction, Transactions},
+    block::{Block, BlockTree, Header, Metadata, Origin, Transaction, Transactions},
     coinbase_puzzle::{CoinbasePuzzle, CoinbaseSolution, EpochChallenge, ProverSolution},
     ledger::helpers::{anchor_block_height, coinbase_reward, coinbase_target, proof_target},
     program::Program,
-    store::{BlockMemory, BlockStorage, BlockStore, ProgramStorage, ProgramStore, TransactionStore, TransitionStore},
+    state_path::StatePath,
+    store::{
+        BlockMemory,
+        BlockStorage,
+        BlockStore,
+        ProgramMemory,
+        ProgramStorage,
+        ProgramStore,
+        TransactionStore,
+        TransitionStore,
+    },
     vm::VM,
-    ProgramMemory,
-    StatePath,
 };
 use console::{
     account::{Address, GraphKey, PrivateKey, Signature, ViewKey},
@@ -713,78 +721,7 @@ impl<N: Network, B: BlockStorage<N>, P: ProgramStorage<N>> Ledger<N, B, P> {
 
     /// Returns a state path for the given commitment.
     pub fn to_state_path(&self, commitment: &Field<N>) -> Result<StatePath<N>> {
-        // Ensure the commitment exists.
-        if !self.contains_commitment(commitment)? {
-            bail!("Commitment '{commitment}' does not exist");
-        }
-
-        // Find the transition that contains the commitment.
-        let transition_id = self.transitions.find_transition_id(commitment)?;
-        // Find the transaction that contains the transition.
-        let transaction_id = match self.transactions.find_transaction_id(&transition_id)? {
-            Some(transaction_id) => transaction_id,
-            None => bail!("The transaction ID for commitment '{commitment}' is not in the ledger"),
-        };
-        // Find the block that contains the transaction.
-        let block_hash = match self.blocks.find_block_hash(&transaction_id)? {
-            Some(block_hash) => block_hash,
-            None => bail!("The block hash for commitment '{commitment}' is not in the ledger"),
-        };
-
-        // Retrieve the transition.
-        let transition = match self.transitions.get_transition(&transition_id)? {
-            Some(transition) => transition,
-            None => bail!("The transition '{transition_id}' for commitment '{commitment}' is not in the ledger"),
-        };
-        // Retrieve the transaction.
-        let transaction = match self.transactions.get_transaction(&transaction_id)? {
-            Some(transaction) => transaction,
-            None => bail!("The transaction '{transaction_id}' for commitment '{commitment}' is not in the ledger"),
-        };
-        // Retrieve the block.
-        let block = match self.blocks.get_block(&block_hash)? {
-            Some(block) => block,
-            None => bail!("The block '{block_hash}' for commitment '{commitment}' is not in the ledger"),
-        };
-
-        // Construct the transition path and transaction leaf.
-        let transition_leaf = transition.to_leaf(commitment, false)?;
-        let transition_path = transition.to_path(&transition_leaf)?;
-
-        // Construct the transaction path and transaction leaf.
-        let transaction_leaf = transaction.to_leaf(transition.id())?;
-        let transaction_path = transaction.to_path(&transaction_leaf)?;
-
-        // Construct the transactions path.
-        let transactions = block.transactions();
-        let transaction_index = transactions.iter().position(|(id, _)| id == &transaction.id()).unwrap();
-        let transactions_path = transactions.to_path(transaction_index, *transaction.id())?;
-
-        // Construct the block header path.
-        let block_header = block.header();
-        let header_root = block_header.to_root()?;
-        let header_leaf = HeaderLeaf::<N>::new(1, block_header.transactions_root());
-        let header_path = block_header.to_path(&header_leaf)?;
-
-        // Construct the state root and block path.
-        let state_root = *self.block_tree.root();
-        let block_path = self.block_tree.prove(block.height() as usize, &block.hash().to_bits_le())?;
-
-        StatePath::new(
-            state_root.into(),
-            block_path,
-            block.hash(),
-            block.previous_hash(),
-            header_root,
-            header_path,
-            header_leaf,
-            transactions_path,
-            transaction.id(),
-            transaction_path,
-            transaction_leaf,
-            transition_path,
-            transition_leaf,
-        )
+        StatePath::new_commitment(&self.block_tree, &self.blocks, &self.transactions, &self.transitions, commitment)
     }
 
     /// Checks the given transaction is well formed and unique.
